requirements.txt
paddlepaddle>=1.8.3
networkx
matplotlib
sphinx
sphinx-rtd-theme
